+++ Magent/Services/ThreadManager+Helpers.swift
@@ -827,6 +827,7 @@ extension Notification.Name {
     static let magentNavigateToThread = Notification.Name("magentNavigateToThread")
     static let magentPullRequestInfoChanged = Notification.Name("magentPullRequestInfoChanged")
     static let magentPromptTOCVisibilityChanged = Notification.Name("magentPromptTOCVisibilityChanged")
+    static let magentSettingsDidChange = Notification.Name("magentSettingsDidChange")
 }
 
 // MARK: - Errors